requirements.txt
pytest
coverage
flask
pyyaml
